third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/function-name-arity-expected.txt

PASS 'use strict'; var f = fn(); f.name = "g"; f.name threw exception of type TypeError.
PASS var f = fn(); f.name = "g"; f.name is "f"
PASS true === true
FAIL 'use strict'; var f = fn(); delete f.name should throw an instance of TypeError
FAIL var f = fn(); delete f.name should be false. Was true.
PASS true === true
PASS 'use strict'; var f = fn(); f.arity = 4; f.arity is 4
PASS var f = fn(); f.arity = 4; f.arity is 4
PASS true === true
PASS 'use strict'; var f = fn(); delete f.arity is true
PASS var f = fn(); delete f.arity is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
